requirements.txt
anyio==3.6.2
black==23.3.0
click==8.1.3
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pycodestyle==2.10.0
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
